Feat: FetchGraphToDepth() to fetch a graph to a given depth
This comes in the context of #5133. It enables Merkledag to fetch
DAGs down to a given depth.

Note that actual usage of depth is expected to be 1, or 2 (and not
an arbitrarily high value), thus I have opted to not complicate things with
branch-pruning optimizations. They can be introduced if they are ever needed
at another point in time.

License: MIT
Signed-off-by: Hector Sanjuan <code@hector.link>