feat: docker build and tag from ci
- build in docker in ci so we can see failures.
- tag and push to dockerhub from ci so we have more control and visibility on the process.

Note, docker workflows run on all branches and no tags by default. You need to opt in to having builds trigger when a git tag is pushed. The filter definition to opt in to tags needs to be present on your job and all dependendent jobs, which is dull. See https://circleci.com/docs/2.0/workflows/#executing-workflows-for-a-git-tag

I've recreated the dockerhub autobuild rules we have currently, but it is worthing taking a moment to review them.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>