feat: add docker tags by post_push hook
Dockerhub allows us to add scripts to a hooks directory in our repo, that allow us to alter the steps of an autobuild!

This PR adds a post_push dockerhub hook that adds additional tags for the stabilize-dht and master branch in the form `<branch>-<short date>-<short commit sha>`, e.g `bifrost-2020-02-26-a44c610` or `master-2020-03-01-18ca5ab`

This is an alternative to #6949 that requires no credentials in CI, but does not ensure our published docker tags are tested.

See #6949 for details

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>