feat: don't put the block back with the other CID version.
We used to do this so we'd provide the block with the correct CID version.
However, we've switched to providing raw multihashes in the DHT so this is no
longer useful.

To not break GC, we _also_ make sure to upgrade CIDs to V1 when garbage
collecting. This ensures that we won't GC a block regardless of how it's
referenced.

The _downside_ is that any blocks previously added using both v0 _and_ v1 won't
be GCed until all pins referencing the block either by v0 or v1 have been
removed.