Added new plugin type: PluginScoreLedger
With a PluginScoreLedger plugin it is now possible to replace the
default BitSwap decision logic with a different one. The plugin
support depends on the experimental BitSwap option `WithScoreLedger`
--- see 80c1a81b63e5 and related commits in the `go-bitswap` project
and also ebf11a6556cc commit in the `go-ipfs-config` project.

A map of score ledgers and new `RegisterScoreLedger` function were
added to `core/node/core.go` in order to support score ledger
injection. The selection of the particular ledger is now possible with
enchanced `OnlineExchange()` function accepting the name of the
ledger (`""` is used for the default behavior). It is configured with
`Experimental.WithScoreLedger` configuration option.

In order to check this modifications and plugin loading two new tests
were added to `test/integration` package: `TestScoreLedgerNotLoaded`
and `TestScoreLedgerLoadStartStop`. To minimize code duplication I've
decided to slightly modify the `core/mock/mock.go` adding new
`NewMockNodeWithConfig()` function there. The code that loads and
injects plugin objects is placed in the `plugins_common_test.go` to
help implementing other plugin tests.