refactor: background=false by default and status
This switches the default to be blocking, just like the old 'pin add`
is. This will make future `pin local add` easier to add in
a backward-compatible manner.

Pinning Services return  only pinned objects by default, which may be
confusing. This adds explicit default and list of available options,
so users are not surprised when non-pinned (eg. queued) pins are ommited
from default output.