Improve error message when running key that locks repo
Fixes Issue #7814

ipfs key commands that open the repo directly will not work when another process has the repo open and locked.  For this reason these commands cannot run when the ipfs daemon is running, and an informative error message should be given to the user.

This is already done for the `ipfs key rotate` command, and should be done for other commands that directly open and lock the repo.  This PR applies the same PreRun check to these commands as is used for `ipfs key rotate`.

While unlikely, it is possible to see the error reported if another process opens the repo at the same moment.  In that case the "someone else has the lock" error is shown.