Make migrations log output to stdout
Migrations were logging to a mix of stdout and stderr.  This was due to the individual migration binaries logging non-error output to stdout, while the migration library (which downloads and executed these migrations) was logging to stderr.  This inconsistency can be confusing.  Also, previous versions of go-ipfs wrote non-error output to stdout. This PR fixes this so that non-error output from the migrations library is written to stdout.

Added test to look for expected log output.