ci: trigger continuation on tags (#8466)
When the config was wrapped in a continuation so that params could be
dynamically generated, this broke tag workflows since there was no
explicit tag filter anymore in config.yml. This attempts to fix that.

I may have gotten this wrong, since this is intrinsically hard to
test. We'll see next time we tag something.