fix: add sharness dir to .dockerignore
Docker doesn't expect the contents of its context to change while it's
building an image, but this happens when building a Docker image while
other sharness tests are running, as they create temp dirs in
test/sharness for storing test artifacts.

This excludes the test/sharness dir from the Docker build
context. It's not included in the resulting image anyway.