Simplifying rcmgr_defaults.go
The goal of this PR is to show an easier set of defaults for resource manager to reason about.
This is an attempt to address https://github.com/ipfs/kubo/issues/9322

The basic idea is:

1. Use these inputs:
 - maxMemory: can be set by user or default 1/8th the system memory
 - maxFD: can be set by user or default 1/2 the system limit
 - maxConns: can be set as the connection manager high water mark or defaults to infinity

2. Only set limits at the system and transient scope, and even there, mostly just focus on memory, FD, and inbound connections.  Ingore outbout connections and stream limits.

3. Apply any limits that libp2p has for its protocols/services.

This PR is not intended to be merged as is.  It's not complete, undoubtedly has syntax errors, I haven't run tests, etc.  It was done as a starting point to communicate specifically on how I think we can simplify the default story.