test: add a test logger to CLI test harness
This logger buffers log events and prints them only if a test fails or
logging has been turned on in code. It does this regardless of the
-verbose flag, so that -verbose can be used without necessarily
showing verbose logs, since it's useful for large test suites since it
streams out individual test progress.

The logger also allows you to create child loggers with additional
prefixes added to each log event, for extra context. Internally, all
loggers in this DAG buffer their own output, and if the test fails or
logging is enabled, the log events trickle back to the root logger
which sorts them all by timestamp and prints them.

This also refactors all the code to use the logger for fatal errors,
which logs the error and then cleanly fails the test, instead of
panicking which aborts the entire package.