Clarity: no user supplied rcmgr limits of 0
The goal of this commit is to make clear that we don't support user-supplied limits of 0 currently.
Apply (https://github.com/libp2p/go-libp2p/blob/master/p2p/host/resource-manager/limit.go#L102 ) overrides any 0 value.  
With the current code, a zero-limit in userSuppliedOverrideLimitConfig will be overridden by the computed default limits.

I want to be able to do:

```
defaultComputedLimitConfig.Apply(userSuppliedOverrideLimitConfig)
```

but that won't override any of the computed defaults.

I'd like to see this fixed, but this at least make clear the situation.