test: make test prereqs like Docker, plugins, etc. opt-in
This is so that by default, you can run tests without having to set up
any of these things and without having to set env vars on the test
runner. This makes it easier to run the CLI tests in your IDE without
having to setup special env vars to skip the Docker, plugin, and fuse tests.