test: port multinode test to Go
This doesn't fully port the test 1:1 because this test isn't actually
doing what it looks like. It looks like it's testing the Websocket
transport with "listentype,ws" IPTB attribute, but that attribute
doesn't exist in ipfs/iptb-plugins, so it does nothing, so the test
actually just runs the same test twice (Yamux disabled). So I've
not ported the no-op Websocket portion of the test.

This also moves equivalent functionality from jbenet/go-random-files
into the testutils package. This just copies the code and modifies it
slightly for better ergonomics.