coreapi/unixfs: in /add, don't setup a MFS root with a mock
It seems to me that this section of the code is unnecessary as a similar setup would be done [in the adder](https://github.com/ipfs/kubo/blob/a7c65184976e8717ac23d7efaa5b0d477ad15deb/core/coreunix/add.go#L86-L101) already.

Additionally that would avoid using a test mock, which is questionable to begin with.