Simplify logic for MFS pinning
MFS pinning logis was returning errors on a channel, only to be logged by a separate goroutine. The errors were not used for any other purpose, so the logic could be significantly simplified by logging the error where it happened instead of returning it on a channel to be logged. This PR removes the unnecessary channel and logging goroutine.

Some other minor simplifications and consistency changes were also made, such as passing a context as the first argument.