Add configuration for bitswap replace WantHave max size
Allow conguration of the bitswap server's replace WantHave with WantBlock maximum block size using the `Internal.Bitswap.ReplaceHasWithBlockMaxSize` config item. This sets the maximum size of a block in bytes up to which we will replace a want-have with a want-block. Setting a size of 0 disables this replacement and means that block sizes are not read for WantHave requests.